+++ bitcoin-fee/bitcoin-fee-estimate-strike/src/main/java/org/tbk/bitcoin/tool/fee/strike/StrikeFeeApiClient.java
@@ -1,0 +1,7 @@
+package org.tbk.bitcoin.tool.fee.strike;
+
+import org.tbk.bitcoin.tool.fee.strike.proto.BlendedFeeEstimateResponse;
+
+public interface StrikeFeeApiClient {
+    BlendedFeeEstimateResponse feeEstimates();
+}